file with collision generation
