+++ Models/Yukawa/yukawaWithCollisionGeneration.py
@@ -12,6 +12,10 @@ import numpy as np
 import WallGo  # Whole package, in particular we get WallGo._initializeInternal()
 from WallGo import Fields, GenericModel, Particle
 
+# Add the Yukawa folder to the path to import YukawaModel
+sys.path.append(str(pathlib.Path(__file__).resolve().parent))
+from yukawa import YukawaModel
+
 # Add the Models folder to the path; need to import the base
 # example template
 modelsBaseDir = pathlib.Path(__file__).resolve().parent.parent
@@ -22,210 +26,6 @@ from wallGoExampleBase import ExampleInputPoint  # pylint: disable=C0411, C0413,
 
 if TYPE_CHECKING:
     import WallGoCollision
-
-
-class YukawaModel(GenericModel):
-    """
-    The Yukawa model, inheriting from WallGo.GenericModel.
-    """
-
-    def __init__(self) -> None:
-        """
-        Initialize the Yukawa model.
-
-        Parameters
-        ----------
-
-        Returns
-        ----------
-        cls: YukawaModel
-            An object of the YukawaModel class.
-        """
-
-        self.modelParameters: dict[str, float] = {}
-
-        # Initialize internal effective potential
-        self.effectivePotential = EffectivePotentialYukawa(self)
-
-        # Create a list of particles relevant for the Boltzmann equations
-        self.defineParticles()
-
-    # ~ GenericModel interface
-    @property
-    def fieldCount(self) -> int:
-        """How many classical background fields"""
-        return 1
-
-    def getEffectivePotential(self) -> "EffectivePotentialYukawa":
-        return self.effectivePotential
-
-    # ~
-
-    def defineParticles(self) -> None:
-        """
-        Define the particles for the model.
-        Note that the particle list only needs to contain the
-        particles that are relevant for the Boltzmann equations.
-        The particles relevant to the effective potential are
-        included independently.
-
-        Parameters
-        ----------
-        None
-
-        Returns
-        ----------
-        None
-        """
-        self.clearParticles()
-
-        # === left fermion ===
-        # The msqVacuum function of an out-of-equilibrium particle must take
-        # a Fields object and return an array of length equal to the number of
-        # points in fields.
-        def psiMsqVacuum(fields: Fields) -> Fields:
-            return (self.modelParameters["mf"] + self.modelParameters["y"] * fields.getField(0)) ** 2
-
-        # The msqDerivative function of an out-of-equilibrium particle must take
-        # a Fields object and return an array with the same shape as fields.
-        def psiMsqDerivative(fields: Fields) -> Fields:  # pylint: disable = W0613
-            return 2 * self.modelParameters["y"] * (self.modelParameters["mf"]
-            + self.modelParameters["y"] * fields.getField(0))
-
-        def psiMsqThermal(T: float) -> float:
-            return 1 / 16 * self.modelParameters["y"] ** 2 * T**2
-
-        psiL = Particle(
-            "psiL",
-            index=1,  # old collision data has top at index 0
-            msqVacuum=psiMsqVacuum,
-            msqDerivative=psiMsqDerivative,
-            msqThermal=psiMsqThermal,
-            statistics="Fermion",
-            totalDOFs=2,
-        )
-        psiR = Particle(
-            "psiR",
-            index=2,  # old collision data has top at index 0
-            msqVacuum=psiMsqVacuum,
-            msqDerivative=psiMsqDerivative,
-            msqThermal=psiMsqThermal,
-            statistics="Fermion",
-            totalDOFs=2,
-        )
-        self.addParticle(psiL)
-        self.addParticle(psiR)
-
-    def calculateLagrangianParameters(
-        self, inputParameters: dict[str, float]
-    ) -> dict[str, float]:
-        """
-        Calculate Lagrangian parameters based on the input parameters.
-
-        Parameters
-        ----------
-        inputParameters: dict[str, float]
-            A dictionary of input parameters for the model.
-
-        Returns
-        ----------
-        modelParameters: dict[str, float]
-            A dictionary of calculated model parameters.
-        """
-        modelParameters = {}
-
-        # Parameters for "phi" field
-        modelParameters = inputParameters
-
-        return modelParameters
-
-    def updateModel(self, newInputParams: dict[str, float]) -> None:
-        """Computes new Lagrangian parameters from given input and caches
-        them internally. These changes automatically propagate to the
-        associated EffectivePotential, particle masses etc.
-        """
-        newParams = self.calculateLagrangianParameters(newInputParams)
-        # Copy to the model dict, do NOT replace the reference.
-        # This way the changes propagate to Veff and particles
-        self.modelParameters.update(newParams)
-
-
-class EffectivePotentialYukawa(WallGo.EffectivePotential):
-    """
-    Effective potential for the Yukawa model.
-
-    This class inherits from the EffectivePotential class and provides the
-    necessary methods for calculating the effective potential.
-    """
-
-    # ~ EffectivePotential interface
-    fieldCount = 1
-    """How many classical background fields"""
-
-    effectivePotentialError = 1e-15
-    """
-    Relative accuracy at which the potential can be computed. Here the potential is
-    polynomial so we can set it to the machine precision.
-    """
-
-    def __init__(self, owningModel: YukawaModel) -> None:
-        """
-        Initialize the EffectivePotentialYukawa.
-        """
-
-        super().__init__()
-
-        assert owningModel is not None, "Invalid model passed to Veff"
-
-        self.owner = owningModel
-        self.modelParameters = self.owner.modelParameters
-
-    def evaluate(
-        self, fields: Fields, temperature: float
-    ) -> float | np.ndarray:
-        """
-        Evaluate the effective potential.
-
-        Parameters
-        ----------
-        fields: Fields
-            The field configuration
-        temperature: float
-            The temperature
-
-        Returns
-        ----------
-        potentialTotal: float | np.ndarray
-            The value of the effective potential
-        """
-        # getting the field from the list of fields (here just of length 1)
-        fields = WallGo.Fields(fields)
-        phi = fields.getField(0)
-
-        # the constant term
-        f0 = -np.pi**2 / 90 * (1 + 4 * 7 / 8) * temperature**4
-
-        # coefficients of the temperature and field dependent terms
-        y = self.modelParameters["y"]
-        mf = self.modelParameters["mf"]
-        sigmaEff = (
-            self.modelParameters["sigma"]
-            + 1 / 24 * (self.modelParameters["gamma"] + 4 * y * mf) * temperature**2
-        )
-        msqEff = (
-            self.modelParameters["msq"]
-            + 1 / 24 * (self.modelParameters["lam"] + 4 * y**2) * temperature**2
-        )
-
-        potentialTotal = (
-            f0
-            + sigmaEff * phi
-            + 1 / 2 * msqEff * phi**2
-            + 1 / 6 * self.modelParameters["gamma"] * phi**3
-            + 1 / 24 * self.modelParameters["lam"] * phi**4
-        )
-
-        return np.array(potentialTotal)
 
 
 class YukawaModelExample(WallGoExampleBase):
@@ -269,6 +69,7 @@ class YukawaModelExample(WallGoExampleBase):
         )
 
         # Add in-equilibrium particles that appear in collision processes
+        # The out-of-equilibrium particles are taken from the definition in the model file
         phiParticle = WallGoCollision.ParticleDescription()
         phiParticle.name = "phi"
         phiParticle.index = 0
@@ -353,8 +154,10 @@ class YukawaModelExample(WallGoExampleBase):
         model.modelParameters, so be careful not to replace that reference here.
         """
 
-        # oldParams = model.modelParameters.copy()
-        model.updateModel(inputParameters)
+        newParams = inputParameters
+        #Copy to the model dict, do NOT replace the reference.
+        # This way the changes propagate to Veff and particles
+        model.modelParameters.update(newParams)
 
     def getBenchmarkPoints(self) -> list[ExampleInputPoint]:
         """
